create a baking rights service abstraction
